refactor: MPT-17825 move audit_paths tests to mirroring test file
After splitting audit_records.py into audit_records.py + audit_paths.py,
move the tests that exercise audit_paths.py functions (get_external_id,
is_valid_path) from tests/cli/plugins/audit_plugin/test_audit_records.py
into a new tests/cli/plugins/audit_plugin/test_audit_paths.py so each
test file mirrors the source module under test.
